Update folder structure for lists
