+++ Lists/task02.py
@@ -1,0 +1,8 @@
+import sys
+
+InputList = sys.argv
+InputList.pop(0)
+
+print("Original List", InputList)
+InputList.reverse()
+print("Original List", InputList)
