+++ Lists/task03.py
@@ -1,0 +1,13 @@
+import sys
+
+checkarg = sys.argv[1]
+mylist = ["5", "banana", "cherry", "test"]
+
+if checkarg in mylist:
+    print("Ist drin")
+else:
+    print("ist nicht drin")
+
+
+
+
