--- task01.py
@@ -1,9 +1,0 @@
-import sys
-
-InputList = sys.argv
-InputList.pop(0)
-max_value = max(InputList)
-
-print(max_value)
-  
-
