--- task02.py
@@ -1,8 +1,0 @@
-import sys
-
-InputList = sys.argv
-InputList.pop(0)
-
-print("Original List", InputList)
-InputList.reverse()
-print("Original List", InputList)
