--- task05.py
@@ -1,6 +1,0 @@
-import sys
-
-InputList = [1,2,3]
-
-result = sum(InputList)
-print(result)
